Extract station_weather_serializers from stations/serializer.py
